+++ app/src/main/java/com/globant/samples/volley/data/repository/UserReposRepository.java
@@ -4,8 +4,8 @@ import android.content.Context;
 
 import com.globant.samples.volley.data.model.repository.GithubUserRepo;
 import com.globant.samples.volley.data.remote.DataManager;
-import com.globant.samples.volley.data.remote.sqlite.room.DataBaseQueries;
-import com.globant.samples.volley.data.remote.sqlite.room.DatabaseCreator;
+import com.globant.samples.volley.data.remote.database.DataBaseQueries;
+import com.globant.samples.volley.data.remote.database.DatabaseCreator;
 import com.globant.samples.volley.injection.qualifier.ApplicationContext;
 
 import java.util.List;
