+++ app/src/main/java/com/globant/samples/volley/injection/component/ApplicationComponent.java
@@ -5,8 +5,9 @@ import android.app.Application;
 import android.content.Context;
 
 import com.globant.samples.volley.data.remote.DataManager;
-import com.globant.samples.volley.data.remote.sqlite.room.DataBaseQueries;
-import com.globant.samples.volley.data.remote.sqlite.room.DatabaseCreator;
+import com.globant.samples.volley.data.remote.database.AppDatabase;
+import com.globant.samples.volley.data.remote.database.DataBaseQueries;
+import com.globant.samples.volley.data.remote.database.DatabaseCreator;
 import com.globant.samples.volley.data.repository.UserRepository;
 import com.globant.samples.volley.data.repository.UserReposRepository;
 import com.globant.samples.volley.injection.module.ApplicationModule;
@@ -40,5 +41,4 @@ public interface ApplicationComponent {
     DatabaseCreator dataBaseCreator();
 
     DataBaseQueries dataBaseQueries();
-
 }
