+++ app/src/main/java/com/globant/samples/volley/injection/module/NetworkModule.java
@@ -5,14 +5,13 @@ import android.content.Context;
 import android.content.SharedPreferences;
 
 import com.globant.samples.volley.data.remote.UserApiService;
+import com.globant.samples.volley.data.remote.database.AppDatabase;
 import com.globant.samples.volley.data.remote.factory.Factory;
-import com.globant.samples.volley.data.remote.sqlite.room.GithubUserDao;
 
 import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
-import io.realm.Realm;
 
 /**
  * Created by miller.barrera.
